refactor: 각 라우트에 해당하는 뷰는 (homeView) home에서만 사용하므로 폴더를 components -> home 으로 이동
