+++ client/src/app/(protected)/search/SearchView/SearchView.style.ts
