+++ client/src/app/(protected)/search/SearchView/stockApi.ts
