+++ client/src/app/(protected)/search/page.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { SearchView } from '@/app/components/SearchView/SearchView';
+import { SearchView } from '@/app/(protected)/search/SearchView/SearchView';
 
 export default function Page() {
   return <SearchView />;
