+++ client/src/app/(public)/main/page.tsx
@@ -1,4 +1,4 @@
-import { MainView } from '@/app/components/MainView/MainView';
+import { MainView } from '@/app/(public)/main/MainView/MainView';
 
 export default function Page() {
   return <MainView />;
